requirements.txt
python-dotenv>=1.0.0
pyrogram>=2.0.0
emoji>=2.8.0
openai>=1.3.0
tgcrypto>=1.2.5  # Pyrogram 的依赖，用于更快的加密
python-logging>=0.4.9 